requirements.txt
numpy>=1.26
scikit-learn>=1.8
Pillow>=11.0
joblib>=1.4
